requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
Django==3.0.8
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
